Remove src from imports, moved out the replacing of describedBy outside
